# job-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Plateforme Emploi",
        "description": "Documentation Swagger des endpoints Admin pour la gestion des utilisateurs, offres et candidatures",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Serveur local"
        }
    ],
    "paths": {
        "/admin/users": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Lister tous les utilisateurs (Admin seulement)",
                "operationId": "7a629ec3ab4a321032bfa425c844d09d",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des utilisateurs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/User"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé - Admin requis"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/users/{id}/role": {
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Modifier le rôle d'un utilisateur (Admin seulement)",
                "operationId": "64fbf4945673fb65321ba32e430bb957",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RoleUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rôle mis à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    },
                    "404": {
                        "description": "Utilisateur non trouvé"
                    },
                    "403": {
                        "description": "Accès refusé - Admin requis"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/jobs": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Lister toutes les offres d'emploi (Admin seulement)",
                "operationId": "b1d00a9ba9f6469e3735190dea65796f",
                "responses": {
                    "200": {
                        "description": "Liste de toutes les offres",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Job"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé - Admin requis"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/applications": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Lister toutes les candidatures (Admin seulement)",
                "operationId": "1645dc31c0de72f14a79a32b57a867f3",
                "responses": {
                    "200": {
                        "description": "Liste des candidatures",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Application"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé - Admin requis"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/statistics": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Récupérer les statistiques globales (Admin seulement)",
                "operationId": "9e3cf48baad7ece8330e837ac83b494b",
                "responses": {
                    "200": {
                        "description": "Statistiques",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Statistics"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé - Admin requis"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/jobs/{id}": {
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Supprimer une offre d'emploi (Admin seulement)",
                "operationId": "dce2d46653d12554027d4fd0cfea2719",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Offre supprimée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Job deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Offre non trouvée"
                    },
                    "403": {
                        "description": "Accès refusé - Admin requis"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/users/{id}": {
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Supprimer un utilisateur (Admin seulement)",
                "operationId": "c3d89b5334aaae1703053279305a83be",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Utilisateur supprimé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Utilisateur non trouvé"
                    },
                    "403": {
                        "description": "Accès refusé - Admin requis"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Statistics": {
                "properties": {
                    "total_users": {
                        "type": "integer",
                        "example": 150
                    },
                    "total_employers": {
                        "type": "integer",
                        "example": 45
                    },
                    "total_candidates": {
                        "type": "integer",
                        "example": 105
                    },
                    "total_jobs": {
                        "type": "integer",
                        "example": 89
                    },
                    "active_jobs": {
                        "type": "integer",
                        "example": 76
                    },
                    "total_applications": {
                        "type": "integer",
                        "example": 234
                    }
                },
                "type": "object"
            },
            "RoleUpdateRequest": {
                "required": [
                    "role"
                ],
                "properties": {
                    "role": {
                        "type": "string",
                        "enum": [
                            "admin",
                            "employer",
                            "candidate"
                        ],
                        "example": "employer"
                    }
                },
                "type": "object"
            },
            "Application": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "job_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "candidate_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "cover_letter": {
                        "type": "string",
                        "example": "Lettre de motivation..."
                    },
                    "resume": {
                        "type": "string",
                        "example": "cv.pdf"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "reviewed",
                            "accepted",
                            "rejected"
                        ],
                        "example": "pending"
                    },
                    "feedback": {
                        "type": "string",
                        "example": "Feedback de l'employeur"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "job": {
                        "$ref": "#/components/schemas/Job"
                    },
                    "candidate": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "title": "User model",
                "required": [
                    "id",
                    "name",
                    "email",
                    "type"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "example": "john@example.com"
                    },
                    "type": {
                        "type": "string",
                        "example": "candidate"
                    },
                    "phone": {
                        "type": "string",
                        "example": "+33123456789"
                    },
                    "bio": {
                        "type": "string",
                        "example": "Développeur fullstack"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-22T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-22T10:00:00Z"
                    }
                },
                "type": "object"
            },
...
            "Job": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "employer_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Développeur Laravel Senior"
                    },
                    "description": {
                        "type": "string",
                        "example": "Description du poste..."
                    },
                    "company": {
                        "type": "string",
                        "example": "Tech Solutions"
                    },
                    "location": {
                        "type": "string",
                        "example": "Paris, France"
                    },
                    "salary": {
                        "type": "number",
                        "format": "float",
                        "example": 55000
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "full_time",
                            "part_time",
                            "contract",
                            "internship"
                        ],
                        "example": "full_time"
                    },
                    "category": {
                        "type": "string",
                        "enum": [
                            "technology",
                            "healthcare",
                            "education",
                            "finance",
                            "other"
                        ],
                        "example": "technology"
                    },
                    "application_deadline": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-12-31"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}